feat: one deck and slides = one canister (#1276)
* feat: init separate canisters for each decks

* feat: sub-canisters owner

* feat: main bucket returns only principal and generate deck bucket type definition

* chore: todo has been solved

* feat: slides in deck bucket canister

* refactor: move deck bucket to separate package / folder

* chore: print out del slide

* chore: doc todo inter canister call

* feat: delete decks and slides upon user delete

* feat: delete decks and slides upon user delete

* feat: set controllers to create canister

* refactor: update actor name

* refactor: bucketIds and canisterId

* feat: deck id should match on update

* fix: handle get error
