+++ studio/canister_ids.json
@@ -5,9 +5,6 @@
   "decks": {
     "ic": "d64ho-qqaaa-aaaah-aaebq-cai"
   },
-  "slides": {
-    "ic": "jkpvl-ziaaa-aaaai-aaafq-cai"
-  },
   "users": {
     "ic": "hic5q-jiaaa-aaaai-aabfq-cai"
   }
