+++ studio/canisters/common/ic.mo
@@ -1,0 +1,59 @@
+// Source: motoko playground
+// https://github.com/dfinity/motoko-playground/blob/3cf294725acf2213ca5e7c01bc59115781ab8ac4/service/pool/IC.mo
+
+// Also update controllers instead of constroller according doc https://sdk.dfinity.org/docs/interface-spec/index.html#ic-interface-overview
+
+module {
+    public type canister_id = Principal;
+    public type canister_settings = {
+        controllers : ?[Principal];
+        freezing_threshold : ?Nat;
+        memory_allocation : ?Nat;
+        compute_allocation : ?Nat;
+    };
+    public type definite_canister_settings = {
+        controllers : ?[Principal];
+        freezing_threshold : Nat;
+        memory_allocation : Nat;
+        compute_allocation : Nat;
+    };
+    public type user_id = Principal;
+    public type wasm_module = Blob;
+
+    public type Self = actor {
+        canister_status : shared query { canister_id : canister_id } -> async {
+            status : { #stopped; #stopping; #running };
+            memory_size : Nat;
+            cycles : Nat;
+            settings : definite_canister_settings;
+            module_hash : ?Blob;
+        };
+        create_canister : shared { settings : ?canister_settings } -> async {
+            canister_id : canister_id;
+        };
+        delete_canister : shared { canister_id : canister_id } -> async ();
+        deposit_cycles : shared { canister_id : canister_id } -> async ();
+        install_code : shared {
+            arg : Blob;
+            wasm_module : wasm_module;
+            mode : { #reinstall; #upgrade; #install };
+            canister_id : canister_id;
+        } -> async ();
+        provisional_create_canister_with_cycles : shared {
+            settings : ?canister_settings;
+            amount : ?Nat;
+        } -> async { canister_id : canister_id };
+        provisional_top_up_canister : shared {
+            canister_id : canister_id;
+            amount : Nat;
+        } -> async ();
+        raw_rand : shared () -> async Blob;
+        start_canister : shared { canister_id : canister_id } -> async ();
+        stop_canister : shared { canister_id : canister_id } -> async ();
+        uninstall_code : shared { canister_id : canister_id } -> async ();
+        update_settings : shared {
+            canister_id : Principal;
+            settings : canister_settings;
+        } -> async ();
+    };
+}
