+++ studio/canisters/deck/deck.data.types.mo
@@ -73,19 +73,4 @@ module {
         data: DeckData;
     };
 
-    public type OwnerDeck = {
-        owner: Principal;
-        deck: Deck;
-    };
-
-    public type ProtectedDeck = {
-        deck: ?OwnerDeck;
-        error: ?Text;
-    };
-
-    public type ProtectedDecks = {
-        decks: [Deck];
-        error: ?Text;
-    };
-
 }
