+++ studio/canisters/deck/slide.data.types.mo
@@ -1,5 +1,4 @@
 import Text "mo:base/Text";
-import Principal "mo:base/Principal";
 import Time "mo:base/Time";
 
 import Types "../common/types";
@@ -7,7 +6,6 @@ import Types "../common/types";
 module {
 
     type SlideId = Types.SlideId;
-    type DeckId = Types.DeckId;
     type Attribute = Types.Attribute;
 
     public type SlideData = {
@@ -24,18 +22,7 @@ module {
 
     public type Slide = {
         slideId: SlideId;
-        deckId: DeckId;
         data: SlideData;
-    };
-
-    public type OwnerSlide = {
-        owner: Principal;
-        slide: Slide;
-    };
-
-    public type ProtectedSlide = {
-        slide: ?OwnerSlide;
-        error: ?Text;
     };
 
 }
