+++ studio/dfx.json
@@ -4,16 +4,12 @@
       "main": "canisters/decks/decks.mo",
       "type": "motoko"
     },
-    "slides": {
-      "main": "canisters/slides/slides.mo",
-      "type": "motoko"
-    },
     "users": {
       "main": "canisters/users/users.mo",
       "type": "motoko"
     },
     "assets": {
-      "dependencies": ["decks", "slides", "users"],
+      "dependencies": ["decks", "users"],
       "frontend": {
         "entrypoint": "www/index.html"
       },
