--- studio/src/app/canisters/slides/slides.did
@@ -1,30 +1,0 @@
-type Time = int;
-type SlideId__1 = text;
-type SlideId = text;
-type SlideData = 
- record {
-   attributes: opt vec Attribute;
-   content: opt text;
-   created_at: opt Time;
-   scope: opt text;
-   template: text;
-   updated_at: opt Time;
- };
-type Slide = 
- record {
-   data: SlideData;
-   deckId: DeckId;
-   slideId: SlideId;
- };
-type DeckId = text;
-type Attribute = 
- record {
-   name: text;
-   value: text;
- };
-service : {
-  del: (SlideId__1) -> (bool);
-  delAdmin: (principal, SlideId__1) -> (bool);
-  get: (SlideId__1) -> (Slide) query;
-  set: (Slide) -> ();
-}
